adjust file structure and add error management in utils
